--- src/test/java/LionTestParam.java
@@ -1,33 +1,0 @@
-import com.example.Feline;
-import com.example.Lion;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-
-import static org.junit.Assert.assertEquals;
-
-@RunWith(Parameterized.class)
-public class LionTestParam {
-    private String sex;
-    private Boolean mane;
-
-    public LionTestParam(String sex, Boolean mane) {
-        this.sex = sex;
-        this.mane = mane;
-    }
-
-    @Parameterized.Parameters
-    public static Object[][] getTestData() {
-        return new Object[][] {
-                {"Самец", true},
-                {"Самка", false}
-        };
-    }
-
-    @Test
-    public void testSex() throws Exception {
-        Feline feline = new Feline();
-        Lion lion = new Lion(sex, feline);
-        assertEquals(mane, lion.doesHaveMane());
-    }
-}
